backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
bcrypt
pyjwt
python-jose